Fixed dragging from textarea in firefox
The bug happened in firefox when user is trying to drag text from textarea. It turns out that in the drag start event, the DataTransfer object is un-readable. Any attempt to read from this object will result in an Error message: Permission denied to access property XXX.

I'm still not sure what caused this. There are 2 potential reasons in mind:
1) According to MDN the permission error is caused by <iframe> violating the same-origin policy. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Property_access_denied
But in our case we aren't using <iframe>
2) There is a "protected mode" for the dnd event. During protected mode, data transfer data will be unavailable. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Property_access_denied
However this mode doesn't happen during drag start. But we are hitting this during drag start.

A quick fix will be adding a try catch block around accessing dataTransfer object.
https://app.asana.com/0/47286296812336/157679098111561