Early return if you drop while not dragging
Similar to https://github.com/Asana/react-dnd-html5-backend/pull/6, patch an invariant crash in ReactDnd. We detect if it's not in a dragging state, but it is trying to call drop action, we catch that and early return.
However I don't feel like I understand the root cause about why this is happening :(